1. compose -> layout 구조로 전환 2. 리스트 표시 로직 추가 3. 프로젝트 설명 추가

+++ app/src/main/java/com/illis/weatherlist/data/model/WeatherDto.kt
@@ -8,7 +8,7 @@ data class WeatherDto(
     val city: City = City()
 ) {
     data class WeatherInfo(
-        val dt: Int = 0,
+        val dt: Long = 0,
         val main: Main = Main(),
         val weather: List<Weather> = listOf(),
         val clouds: Clouds = Clouds(),
